"Engage!" game board (redesigned)
This game board for the "engage!" game (http://schreibman.eu/engage/) was redesigned by Isabelle Reiß as part of her DH internship at the Institute of European History Mainz in 2019. You are welcome to use the "engage!" board game and all other materials we provide free of charge and to distribute as many copies as you like. However, we ask you to credit us and to adhere to our CC-BY-SA license.